Move common functions to module
