+++ tuning/scanners.py
@@ -2,9 +2,11 @@
 
 from abc import ABC, abstractmethod
 from copy import copy
-from typing import List, Union
+from typing import Dict, List, Union
 
 import numpy as np
+
+from tuning import scanners
 
 # One dimensional scanners
 
@@ -262,3 +264,38 @@ class SpiralScanner(Scanner2D):
 
     @property
     def index(self) -> List[int]: self.index = [self.n_arm, self.step]
+
+def read_cell(excel_file, polarimeter: str, test: str, mode: str) -> Union[Scanner2D, Scanner1D]:
+    from ast import literal_eval
+
+    row = excel_file[polarimeter]
+    if test != "Offset":
+        test += mode
+    scanner_class = getattr(scanners, row[(test, "Scanner")])
+    arguments_str = row[(test, "Arguments")]
+    arguments = list(map(literal_eval, arguments_str.split(";")))
+    for i in range(len(arguments)):
+        if isinstance(arguments[i], list):
+            arguments[i] = np.asarray(arguments[i], dtype=float)
+    if test == "Offset":
+        return scanner_class(*arguments, label="offset")
+    else:
+        return scanner_class(*arguments, x_label="idrain", y_label="offset")
+
+def read_excel(filename: str, dummy_polarimeter: bool = False, open_loop: bool = False) -> Dict[str, Dict[str, Scanner2D]]:
+    import pandas as pd
+
+    excel_file = pd.read_excel(filename, header=(0, 1), index_col=0).to_dict(orient="index")
+    scanners = {}
+    if open_loop:
+        mode = " open loop"
+    else:
+        mode = " closed loop"
+    for polarimeter in set(excel_file) - {"DUMMY"}: # Iterate over all polarimeters except the DUMMY one
+        scanners[polarimeter] = {}
+        for test in "HA1", "HA2", "HA3", "HB1", "HB2", "HB3", "Offset":
+            if dummy_polarimeter:
+                scanners[polarimeter][test] = read_cell(excel_file, "DUMMY", test, mode)
+            else:
+                scanners[polarimeter][test] = read_cell(excel_file, polarimeter, test, mode)
+    return scanners
